todo/python/002-区划信息/script/requirements.txt

requests_html
lxml_html_clean
html2text
mysql-connector-python

